@@ -1,27 +1,27 @@
 {
-  "_from": "qs@6.9.6",
-  "_id": "qs@6.9.6",
+  "_from": "qs@6.9.7",
+  "_id": "qs@6.9.7",
   "_inBundle": false,
-  "_integrity": "sha512-TIRk4aqYLNoJUbd+g2lEdz5kLWIuTMRagAXxl78Q0RiVjAOugHmeKNGdd3cwo/ktpf9aL9epCfFqWDEKysUlLQ==",
+  "_integrity": "sha512-IhMFgUmuNpyRfxA90umL7ByLlgRXu6tIfKPpF5TmcfRLlLCckfP/g3IQmju6jjpu+Hh8rA+2p6A27ZSPOOHdKw==",
   "_location": "/body-parser/qs",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "qs@6.9.6",
+    "raw": "qs@6.9.7",
     "name": "qs",
     "escapedName": "qs",
-    "rawSpec": "6.9.6",
+    "rawSpec": "6.9.7",
     "saveSpec": null,
-    "fetchSpec": "6.9.6"
+    "fetchSpec": "6.9.7"
   },
   "_requiredBy": [
     "/body-parser"
   ],
-  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.9.6.tgz",
-  "_shasum": "26ed3c8243a431b2924aca84cc90471f35d5a0ee",
-  "_spec": "qs@6.9.6",
-  "_where": "/home/kilatex/Dev-Projects/Vue/LearningVue/vue-project/node_modules/body-parser",
+  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.9.7.tgz",
+  "_shasum": "4610846871485e1e048f44ae3b94033f0e675afe",
+  "_spec": "qs@6.9.7",
+  "_where": "/home/kilatex/Dev-Projects/Vue/Templates/vue-project/node_modules/body-parser",
   "bugs": {
     "url": "https://github.com/ljharb/qs/issues"
   },
@@ -33,26 +33,26 @@
       "url": "http://ljharb.codes"
     }
   ],
-  "dependencies": {},
   "deprecated": false,
   "description": "A querystring parser that supports nesting and arrays, with a depth limit",
   "devDependencies": {
-    "@ljharb/eslint-config": "^17.3.0",
-    "aud": "^1.1.3",
+    "@ljharb/eslint-config": "^20.1.0",
+    "aud": "^1.1.5",
     "browserify": "^16.5.2",
     "eclint": "^2.8.1",
-    "eslint": "^7.17.0",
+    "eslint": "^8.6.0",
     "evalmd": "^0.0.19",
     "for-each": "^0.3.3",
-    "has-symbols": "^1.0.1",
+    "has-symbols": "^1.0.2",
     "iconv-lite": "^0.5.1",
+    "in-publish": "^2.0.1",
     "mkdirp": "^0.5.5",
     "nyc": "^10.3.2",
-    "object-inspect": "^1.9.0",
+    "object-inspect": "^1.12.0",
     "qs-iconv": "^1.0.4",
-    "safe-publish-latest": "^1.1.4",
+    "safe-publish-latest": "^2.0.0",
     "safer-buffer": "^2.1.2",
-    "tape": "^5.1.1"
+    "tape": "^5.4.0"
   },
   "engines": {
     "node": ">=0.6"
@@ -84,14 +84,15 @@
   },
   "scripts": {
     "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js",
-    "lint": "eslint lib/*.js test/*.js",
-    "postlint": "eclint check * lib/* test/*",
+    "lint": "eslint .",
+    "postlint": "eclint check * lib/* test/* !dist/*",
     "posttest": "aud --production",
-    "prepublish": "safe-publish-latest && npm run dist",
+    "prepublish": "not-in-publish || npm run prepublishOnly",
+    "prepublishOnly": "safe-publish-latest && npm run dist",
     "pretest": "npm run --silent readme && npm run --silent lint",
     "readme": "evalmd README.md",
     "test": "npm run tests-only",
     "tests-only": "nyc tape 'test/**/*.js'"
   },
-  "version": "6.9.6"
+  "version": "6.9.7"
 }
