@@ -1,27 +1,27 @@
 {
-  "_from": "cookie@0.4.1",
-  "_id": "cookie@0.4.1",
+  "_from": "cookie@0.4.2",
+  "_id": "cookie@0.4.2",
   "_inBundle": false,
-  "_integrity": "sha512-ZwrFkGJxUR3EIoXtO+yVE69Eb7KlixbaeAWfBQB9vVsNn/o+Yw69gBWSSDK825hQNdN+wF8zELf3dFNl/kxkUA==",
+  "_integrity": "sha512-aSWTXFzaKWkvHO1Ny/s+ePFpvKsPnjc551iI41v3ny/ow6tBG5Vd+FuqGNhh1LxOmVzOlGUriIlOaokOvhaStA==",
   "_location": "/cookie",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "cookie@0.4.1",
+    "raw": "cookie@0.4.2",
     "name": "cookie",
     "escapedName": "cookie",
-    "rawSpec": "0.4.1",
+    "rawSpec": "0.4.2",
     "saveSpec": null,
-    "fetchSpec": "0.4.1"
+    "fetchSpec": "0.4.2"
   },
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/cookie/-/cookie-0.4.1.tgz",
-  "_shasum": "afd713fe26ebd21ba95ceb61f9a8116e50a537d1",
-  "_spec": "cookie@0.4.1",
-  "_where": "/home/kilatex/Dev-Projects/Vue/LearningVue/vue-project/node_modules/express",
+  "_resolved": "https://registry.npmjs.org/cookie/-/cookie-0.4.2.tgz",
+  "_shasum": "0e41f24de5ecf317947c82fc789e06a884824432",
+  "_spec": "cookie@0.4.2",
+  "_where": "/home/kilatex/Dev-Projects/Vue/Templates/vue-project/node_modules/express",
   "author": {
     "name": "Roman Shtylman",
     "email": "shtylman@gmail.com"
@@ -41,10 +41,11 @@
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "6.8.0",
-    "eslint-plugin-markdown": "1.0.2",
-    "mocha": "7.1.1",
-    "nyc": "15.0.1"
+    "eslint": "7.32.0",
+    "eslint-plugin-markdown": "2.2.1",
+    "mocha": "9.2.0",
+    "nyc": "15.1.0",
+    "top-sites": "1.1.85"
   },
   "engines": {
     "node": ">= 0.6"
@@ -68,11 +69,12 @@
   },
   "scripts": {
     "bench": "node benchmark/index.js",
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks --ui qunit test/",
-    "test-ci": "nyc --reporter=text npm test",
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "update-bench": "node scripts/update-benchmark.js",
     "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "0.4.1"
+  "version": "0.4.2"
 }
