@@ -35,33 +37,44 @@
       "name": "Michał Sajnóg",
       "email": "msajnog93@gmail.com",
       "url": "https://github.com/michalsnik"
+    },
+    {
+      "name": "Yosuke Ota",
+      "url": "https://github.com/ota-meshi"
     }
   ],
   "dependencies": {
+    "eslint-utils": "^2.1.0",
     "natural-compare": "^1.4.0",
-    "semver": "^5.6.0",
-    "vue-eslint-parser": "^7.0.0"
+    "semver": "^6.3.0",
+    "vue-eslint-parser": "^7.10.0"
   },
   "deprecated": false,
   "description": "Official ESLint plugin for Vue.js",
   "devDependencies": {
-    "@types/node": "^4.2.16",
-    "@typescript-eslint/parser": "^2.6.1",
-    "acorn": "^7.1.0",
-    "babel-eslint": "^10.0.2",
-    "chai": "^4.1.0",
-    "eslint": "^6.0.0",
-    "eslint-plugin-eslint-plugin": "^2.0.1",
-    "eslint-plugin-import": "^2.18.2",
+    "@types/eslint": "^7.28.1",
+    "@types/eslint-visitor-keys": "^1.0.0",
+    "@types/natural-compare": "^1.4.0",
+    "@types/node": "^13.13.5",
+    "@types/semver": "^7.2.0",
+    "@typescript-eslint/parser": "^5.0.0-0",
+    "@vuepress/plugin-pwa": "^1.4.1",
+    "acorn": "^8.5.0",
+    "env-cmd": "^10.1.0",
+    "eslint": "^8.0.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-eslint-plugin": "^3.5.3",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-vue": "file:.",
-    "eslint-plugin-vue-libs": "^4.0.0",
-    "eslint4b": "^6.6.0",
-    "lodash": "^4.17.4",
-    "mocha": "^5.2.0",
-    "nyc": "^12.0.2",
-    "typescript": "^3.5.2",
-    "vue-eslint-editor": "^0.1.4",
-    "vuepress": "^0.14.5"
+    "espree": "^9.0.0",
+    "lodash": "^4.17.21",
+    "mocha": "^7.1.2",
+    "nyc": "^15.1.0",
+    "prettier": "^2.4.1",
+    "typescript": "^4.4.3",
+    "vue-eslint-editor": "^1.1.0",
+    "vuepress": "^1.8.2"
   },
   "engines": {
     "node": ">=8.10"
@@ -82,24 +95,30 @@
   "main": "lib/index.js",
   "name": "eslint-plugin-vue",
   "peerDependencies": {
-    "eslint": "^5.0.0 || ^6.0.0"
+    "eslint": "^6.2.0 || ^7.0.0 || ^8.0.0"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/vuejs/eslint-plugin-vue.git"
   },
   "scripts": {
-    "debug": "mocha --inspect-brk \"tests/lib/**/*.js\" --reporter dot --timeout 60000",
+    "cover": "npm run cover:test && npm run cover:report",
+    "cover:report": "nyc report --reporter=html",
+    "cover:test": "nyc npm run test:base -- --timeout 60000",
+    "debug": "mocha --inspect \"tests/lib/**/*.js\" --reporter dot --timeout 60000",
     "docs:build": "vuepress build docs",
     "docs:watch": "vuepress dev docs",
     "lint": "eslint . --rulesdir eslint-internal-rules",
-    "pretest": "npm run lint",
-    "preversion": "npm test && npm run update && git add .",
+    "lint:fix": "eslint . --rulesdir eslint-internal-rules --fix",
+    "new": "node tools/new-rule.js",
+    "predocs:build": "npm run update",
+    "preversion": "npm test && git add .",
     "start": "npm run test:base -- --watch --growl",
     "test": "nyc npm run test:base -- \"tests/integrations/*.js\" --timeout 60000",
     "test:base": "mocha \"tests/lib/**/*.js\" --reporter dot",
+    "tsc": "tsc",
     "update": "node ./tools/update.js",
-    "version": "npm run lint -- --fix && git add ."
+    "version": "env-cmd -e version npm run update && npm run lint -- --fix && git add ."
   },
-  "version": "6.2.2"
+  "version": "7.20.0"
 }
