@@ -1,23 +1,10 @@
 <template>
-  <div id="app">
-    <div id="nav">
-      <router-link to="/">Home</router-link> |
-      <router-link to="/about">About</router-link>
-
-      <router-view></router-view>
-    </div>
+  <div id="nav">
+    <router-link to="/">Home</router-link> |
+    <router-link to="/about">About</router-link>
   </div>
+  <router-view/>
 </template>
-
-
-<script>
-export default {
-  name: 'app',
-  components:{
-
-  }
-}
-</script>
 
 <style>
 #app {
